multitask-learning/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv

torch
transformers
sentencepiece
scikit-learn
pandas
numpy
datasets
evaluate
psycopg2-binary
text2emotion
nltk
Sastrawi